Full refactor of TelegramBotService: split into modular services for translations and AI context
